practicas/practica_final/requirements.txt
mcp>=1.2.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langchain-chroma>=0.1.4
langchain-text-splitters>=0.3.0
chromadb>=0.5.0
python-dotenv>=1.0.0
